fix: rewrite button component to ensure correct styling + moved cn to its own helper function file
