+++ components/shared/button/Button.tsx
@@ -2,10 +2,10 @@ import { Slot } from "@radix-ui/react-slot"
 import { cva, type VariantProps } from "class-variance-authority"
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/helpers/helper.cn"
 
 const buttonVariants = cva(
-  `inline-flex border text-typo-body-sm text-foreground bg-background shadow-button rounded-full items-center justify-center
+  `inline-flex border border-foreground text-typo-body-sm text-foreground bg-background shadow-button rounded-full items-center justify-center
   whitespace-nowrap font-medium font-headline focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring
   disabled:pointer-events-none disabled:opacity-50 hover:translate-x-[1px] hover:translate-y-[1px] transition
   hover:shadow-buttonHover active:translate-x-[4px] active:translate-y-[4px] active:shadow-none `,
@@ -60,9 +60,7 @@ export interface ButtonProps
 const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
   ({ className, variant, size, asChild = false, ...props }, ref) => {
     const Comp = asChild ? Slot : "button"
-    // TODO: This sucks. @ThomasGross what to do?
-    const classes = cn(buttonVariants({ variant, size, className }))
-    return <Comp className={classes} ref={ref} {...props} />
+    return <Comp className={cn(buttonVariants({ variant, size }), className)} ref={ref} {...props} />
   }
 )
