+++ components/shared/searchInput/SearchInput.tsx
@@ -4,7 +4,7 @@ import { useRouter, useSearchParams } from "next/navigation"
 import React from "react"
 import { useEffect, useRef } from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/helpers/helper.cn"
 
 import Icon from "../icon/Icon"
