+++ components/shared/sheet/LoginSheet.tsx
@@ -5,7 +5,7 @@ import { Cross2Icon } from "@radix-ui/react-icons"
 import { cva, type VariantProps } from "class-variance-authority"
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/helpers/helper.cn"
 
 const Sheet = SheetPrimitive.Root
